requirements.txt
langchain
langgraph
ragas
chromadb
streamlit
pydantic
python-dotenv
pytest
pypdf
langchain-chroma
fastapi